Fixed rsa automatic generation, updated from sharpssh to sshnet
*Fixed issue #3 regarding the automatic ssl key generation. This
required introducing an external Utilities project, which provides an
OpenSSL layer for key generation (had some trouble using the ones
found).

*Also the core library of ssh was changed from outdated SharpSSH to
SSH.NET which is currently active, and simpler.

*Added automatic cleanup of closed connections, and when a socket error
occurs (such as a timeout) it closes and displays a message to the user
automatically.

*Added Installer project, hopefully, this is where the build in each
tag will be placed.
*Added Test project, still blank though.
*Added missing Windows7Api code pack dlls
*Now we are using the private key as base file, instead of setting the
public key and then assuming the private is on the same but without the
*.pub extension.